We need some system level tests with full coverage.
