+++ lib/cli.js
@@ -7,24 +7,29 @@ var cli_punish = require("./cli/punish");
 var cli_auth = require("./cli/auth");
 var cli_init = require("./cli/init");
 var pkg = require("./../package");
-var no_args = function no_args() {
-    return !process.argv.slice(2).length;
+var no_args = function no_args(argv) {
+    return !argv.slice(2).length;
 };
-var configure = function configure() {
+var log_additional_help = function log_additional_help() {
+    console.log("  Command specific help:");
+    console.log();
+    console.log("    {cmd} -h, --help");
+    console.log();
+};
+var sub_modules = function sub_modules() {
+    return [cli_punish, cli_auth, cli_init];
+};
+var bind_sub_module = function bind_sub_module(cli_sub_mod) {
+    cli_sub_mod.create(cli);
+};
+var configure = function configure(argv) {
     cli.version(pkg.version);
-    cli.on("--help", function () {
-        console.log("  Command specific help:");
-        console.log();
-        console.log("    {cmd} -h, --help");
-        console.log();
-    });
-    _.each([cli_punish, cli_auth, cli_init], function (command) {
-        command.create(cli);
-    });
-    if (no_args()) cli.outputHelp();
+    _.each(sub_modules(), bind_sub_module);
+    cli.on("--help", log_additional_help);
+    if (no_args(argv)) cli.outputHelp();
 };
 var interpret = function interpret(argv) {
-    return configure(), cli.parse(argv);
+    return configure(argv), cli.parse(argv);
 };
 module.exports = {
     interpret: interpret
