+++ lib/index.js
@@ -5,7 +5,5 @@ var extend = require("extend");
 var plugins = require("./plugin");
 var util = require("./util");
 var logger = require("./logger");
-// TODO: flush out
-// TODO: require in config now?!
 var library = extend({}, plugins, util, { logger: logger });
 module.exports = library;
