+++ lib/plugin.js
@@ -246,51 +246,6 @@ var normalize_paths = function normalize_paths(issues) {
         }
     });
 };
-var set_ignore_list = function set_ignore_list(plugin_config, base) {
-    var list = _.get(plugin_config, "ignore", []);
-    if (_.isString(list)) list = [list];
-    if (_.isEmpty(list)) {
-        _.set(plugin_config, "ignore", base);
-    } else if (is_array(list)) {
-        _.set(plugin_config, "ignore", _.uniq(list.concat(base)));
-    }
-};
-var set_allow_list = function set_allow_list(plugin_config, base) {
-    if (!_.isEmpty(base)) {
-        if (_.isString(base)) base = [base];
-        _.set(plugin_config, "allow", base);
-    } else {
-        var list = _.get(plugin_config, "allow", []);
-        if (_.isString(list)) list = [list];
-        _.set(plugin_config, "allow", list);
-    }
-};
-var get_plugin_config = function get_plugin_config(name, config) {
-    var plugin_config = _.get(config, name, {});
-    var vile_ignore = _.get(config, "vile.ignore", []);
-    var vile_allow = _.get(config, "vile.allow", []);
-    set_ignore_list(plugin_config, vile_ignore);
-    set_allow_list(plugin_config, vile_allow);
-    return plugin_config;
-};
-var ping_parent = function ping_parent(process) {
-    return process.send("");
-};
-var handle_worker_request = function handle_worker_request(data) {
-    var plugins = data.plugins;
-    var config = data.config;
-    Bluebird.map(plugins, function (plugin) {
-        var name = plugin.replace("vile-", "");
-        var plugin_config = get_plugin_config(name, config);
-        return run_plugin(name, plugin_config).catch(function (err) {
-            console.log(); // newline because spinner is running
-            log.error(err.stack || err);
-            process.exit(1);
-        });
-    }).then(_.flatten).then(function (issues) {
-        return process.send(issues);
-    });
-};
 var check_for_uninstalled_plugins = function check_for_uninstalled_plugins(allowed, plugins) {
     var errors = false;
     _.each(allowed, function (name) {
@@ -355,57 +310,53 @@ var execute_plugins = function execute_plugins() {
     return function (plugins) {
         return new Bluebird(function (resolve, reject) {
             check_for_uninstalled_plugins(allowed, plugins);
-            if (cluster.isMaster) {
-                (function () {
-                    // TODO: own method
-                    if (allowed.length > 0) {
-                        plugins = _.filter(plugins, function (p) {
-                            return _.some(allowed, function (a) {
-                                return p.replace("vile-", "") == a;
-                            });
-                        });
-                    }
-                    var spin = void 0;
-                    var workers = {};
-                    var plugin_count = plugins.length;
-                    var concurrency = os.cpus().length || 1;
-                    cluster.on("fork", function (worker) {
-                        if (spin) spin.stop(true);
-                        log.info(workers[worker.id] + ":start " + ("(" + worker.id + "/" + plugin_count + ")"));
-                        if (spin) spin.start();
+            cluster.setupMaster({
+                exec: path.join(__dirname, "plugin", "worker.js")
+            });
+            // TODO: own method
+            if (allowed.length > 0) {
+                plugins = _.filter(plugins, function (p) {
+                    return _.some(allowed, function (a) {
+                        return p.replace("vile-", "") == a;
                     });
-                    if (opts.spinner && opts.format != "json") {
-                        spin = new Spinner("PUNISH");
-                        spin.setSpinnerDelay(60);
-                        spin.start();
-                    }
-                    Bluebird.map(plugins, function (plugin) {
-                        var worker = cluster.fork();
-                        workers[worker.id] = plugin.replace("vile-", "");
-                        return run_plugins_in_fork([plugin], config, worker).then(function (issues) {
-                            return normalize_paths(issues), issues;
-                        }).catch(function (err) {
-                            if (spin) spin.stop(true);
-                            log.error(err.stack || err);
-                            reject(err);
-                        });
-                    }, { concurrency: concurrency }).then(_.flatten).then(function (issues) {
-                        if (spin) spin.stop(true);
-                        if (!_.isEmpty(opts.combine)) {
-                            issues = combine_paths(opts.combine, issues);
-                        }
-                        if (opts.format == "syntastic") {
-                            log_syntastic_applicable_messages(issues);
-                        } else if (opts.format != "json") {
-                            log_issue_messages(issues);
-                        }
-                        resolve(issues);
-                    });
-                })();
-            } else {
-                process.on("message", handle_worker_request);
-                ping_parent(process);
+                });
             }
+            var spin = void 0;
+            var workers = {};
+            var plugin_count = plugins.length;
+            var concurrency = os.cpus().length || 1;
+            cluster.on("fork", function (worker) {
+                if (spin) spin.stop(true);
+                log.info(workers[worker.id] + ":start " + ("(" + worker.id + "/" + plugin_count + ")"));
+                if (spin) spin.start();
+            });
+            if (opts.spinner && opts.format != "json") {
+                spin = new Spinner("PUNISH");
+                spin.setSpinnerDelay(60);
+                spin.start();
+            }
+            Bluebird.map(plugins, function (plugin) {
+                var worker = cluster.fork();
+                workers[worker.id] = plugin.replace("vile-", "");
+                return run_plugins_in_fork([plugin], config, worker).then(function (issues) {
+                    return normalize_paths(issues), issues;
+                }).catch(function (err) {
+                    if (spin) spin.stop(true);
+                    log.error(err.stack || err);
+                    reject(err);
+                });
+            }, { concurrency: concurrency }).then(_.flatten).then(function (issues) {
+                if (spin) spin.stop(true);
+                if (!_.isEmpty(opts.combine)) {
+                    issues = combine_paths(opts.combine, issues);
+                }
+                if (opts.format == "syntastic") {
+                    log_syntastic_applicable_messages(issues);
+                } else if (opts.format != "json") {
+                    log_issue_messages(issues);
+                }
+                resolve(issues);
+            });
         });
     };
 };
