+++ package.json
@@ -28,7 +28,9 @@
     "mocha": "mocha",
     "ncu": "ncu -a",
     "release-it": "release-it",
-    "test": "node bin/test",
+    "test": "node bin/test-all",
+    "test-system": "node bin/test-system",
+    "test-non-system": "node bin/test-non-system",
     "test-cov": "node bin/test-cov",
     "test-ci-build": "node bin/test-ci-build",
     "tsc": "tsc",
@@ -54,13 +56,13 @@
     "tool"
   ],
   "dependencies": {
-    "bluebird": "^3.4.3",
+    "bluebird": "^3.4.6",
     "cli-spinner": "0.2.5",
     "commander": "^2.9.0",
     "extend": "^3.0.0",
     "git-diff-tree": "^1.0.0",
     "ignore-file": "^1.1.2",
-    "inquirer": "^1.1.2",
+    "inquirer": "^1.1.3",
     "js-yaml": "^3.6.1",
     "language-detect": "^1.1.0",
     "linez": "^4.1.1",
@@ -94,17 +96,18 @@
     "chai-as-promised": "^5.3.0",
     "coffee-script": "^1.10.0",
     "istanbul": "^0.4.5",
-    "mimus": "^0.2.2",
+    "mimus": "^0.2.3",
     "mocha": "^3.0.2",
     "npm-check-updates": "^2.8.0",
+    "nyc": "^8.1.0",
     "pryjs": "^1.0.2",
     "release-it": "^2.4.3",
-    "retire": "^1.1.5",
+    "retire": "^1.2.10",
     "shelljs": "^0.7.4",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
     "typescript": "^1.8.10",
-    "typescript-eslint-parser": "0.1.3",
+    "typescript-eslint-parser": "0.2.0",
     "wrench": "^1.5.9"
   }
 }
