+++ test/fixtures/system/async/.vile.yml
@@ -1,0 +1,4 @@
+vile:
+  ignore:
+    - node_modules
+    - npm-debug.log
