+++ test/helpers/system.coffee
@@ -1,0 +1,4 @@
+exec = ->
+
+module.exports =
+  exec: exec
