+++ test/integration/plugin.coffee
@@ -1,0 +1,2 @@
+
+describe "plugin", ->
